refactor: Remove unused page components and update file structure
