--- src/app/(admin)/admin/users/page.tsx
@@ -1,7 +1,0 @@
-export default function Users() {
-  return (
-    <main className="flex min-h-screen flex-col items-center justify-between p-24">
-      Users
-    </main>
-  );
-}
