--- src/app/dashboard/participant/page.tsx
@@ -1,3 +1,0 @@
-export default function page() {
-  return <div>page</div>;
-}
