+++ src/components/Dashboard/DevPostSubmission.tsx
@@ -1,0 +1,9 @@
+import Card from "./Card";
+
+export default function DevPostSubmission() {
+  return (
+    <Card>
+      <div>DevPostSubmission</div>
+    </Card>
+  );
+}
